// src/api/generated/models/SedationistDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RoleType } from './RoleType';
export type SedationistDto = {
    id?: string;
    address?: string | null;
    town?: string | null;
    country?: string | null;
    postCode?: string | null;
    roleType?: RoleType;
    userId?: string;
    email?: string | null;
    createdDateTime?: string;
    firstName?: string | null;
    lastName?: string | null;
    fullName?: string | null;
    phoneNumber?: string | null;
    notes?: string | null;
};