// src/api/generated/models/AvailabilityStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum AvailabilityStatus {
    UNAVAILABLE = 'Unavailable',
    ON_LEAVE = 'OnLeave',
    MEETING = 'Meeting',
    ATTENDING_WORKSHOP = 'AttendingWorkshop',
    CANCELLED = 'Cancelled',
    RESCHEDULED = 'Rescheduled',
    PERSONAL_REASON = 'Personal Reason',
}